+++ test/test_go.rb
@@ -430,6 +430,10 @@ class TestGoFZF < TestBase
     test.call 'f3', 'f3'
     test.call 'f2,f4', 'f2', 'f2'
     test.call 'f2,f4', 'f4', 'f4'
+    test.call 'alt-/', [:Escape, :/]
+    %w[f5 f6 f7 f8 f9 f10].each do |key|
+      test.call 'f5,f6,f7,f8,f9,f10', key, key
+    end
     test.call '@', '@'
   end
 
@@ -771,6 +775,13 @@ class TestGoFZF < TestBase
     assert_equal %w[4 5 6 9], readonce.split($/)
   end
 
+  def test_bind_print_query
+    tmux.send_keys "seq 1 1000 | #{fzf '-m --bind=ctrl-j:print-query'}", :Enter
+    tmux.until { |lines| lines[-2].end_with? '/1000' }
+    tmux.send_keys 'print-my-query', 'C-j'
+    assert_equal %w[print-my-query], readonce.split($/)
+  end
+
   def test_long_line
     data = '.' * 256 * 1024
     File.open(tempname, 'w') do |f|
@@ -858,20 +869,34 @@ class TestGoFZF < TestBase
 
   def test_execute
     output = '/tmp/fzf-test-execute'
-    opts = %[--bind \\"alt-a:execute(echo '[{}]' >> #{output}),alt-b:execute[echo '({}), ({})' >> #{output}],C:execute:echo '({}), [{}], @{}@' >> #{output}\\"]
-    tmux.send_keys "seq 100 | #{fzf opts}", :Enter
-    tmux.until { |lines| lines[-2].include? '100/100' }
-    tmux.send_keys :Escape, :a, :Escape, :a
+    opts = %[--bind \\"alt-a:execute(echo [{}] >> #{output}),alt-b:execute[echo /{}{}/ >> #{output}],C:execute:echo /{}{}{}/ >> #{output}\\"]
+    wait = lambda { |exp| tmux.until { |lines| lines[-2].include? exp } }
+    writelines tempname, %w[foo'bar foo"bar foo$bar]
+    tmux.send_keys "cat #{tempname} | #{fzf opts}; sync", :Enter
+    wait['3/3']
+    tmux.send_keys :Escape, :a
+    wait['/3']
+    tmux.send_keys :Escape, :a
+    wait['/3']
     tmux.send_keys :Up
-    tmux.send_keys :Escape, :b, :Escape, :b
+    tmux.send_keys :Escape, :b
+    wait['/3']
+    tmux.send_keys :Escape, :b
+    wait['/3']
     tmux.send_keys :Up
     tmux.send_keys :C
-    tmux.send_keys 'foobar'
-    tmux.until { |lines| lines[-2].include? '0/100' }
-    tmux.send_keys :Escape, :a, :Escape, :b, :Escape, :c
+    wait['3/3']
+    tmux.send_keys 'barfoo'
+    wait['0/3']
+    tmux.send_keys :Escape, :a
+    wait['/3']
+    tmux.send_keys :Escape, :b
+    wait['/3']
     tmux.send_keys :Enter
     readonce
-    assert_equal ['["1"]', '["1"]', '("2"), ("2")', '("2"), ("2")', '("3"), ["3"], @"3"@'],
+    assert_equal %w[[foo'bar] [foo'bar]
+                    /foo"barfoo"bar/ /foo"barfoo"bar/
+                    /foo$barfoo$barfoo$bar/],
       File.readlines(output).map(&:chomp)
   ensure
     File.unlink output rescue nil
@@ -879,21 +904,24 @@ class TestGoFZF < TestBase
 
   def test_execute_multi
     output = '/tmp/fzf-test-execute-multi'
-    opts = %[--multi --bind \\"alt-a:execute-multi(echo '[{}], @{}@' >> #{output}; sync)\\"]
-    tmux.send_keys "seq 100 | #{fzf opts}", :Enter
-    tmux.until { |lines| lines[-2].include? '100/100' }
+    opts = %[--multi --bind \\"alt-a:execute-multi(echo {}/{} >> #{output}; sync)\\"]
+    writelines tempname, %w[foo'bar foo"bar foo$bar foobar]
+    tmux.send_keys "cat #{tempname} | #{fzf opts}", :Enter
+    tmux.until { |lines| lines[-2].include? '4/4' }
     tmux.send_keys :Escape, :a
-    tmux.until { |lines| lines[-2].include? '/100' }
+    tmux.until { |lines| lines[-2].include? '/4' }
     tmux.send_keys :BTab, :BTab, :BTab
     tmux.send_keys :Escape, :a
-    tmux.until { |lines| lines[-2].include? '/100' }
+    tmux.until { |lines| lines[-2].include? '/4' }
     tmux.send_keys :Tab, :Tab
     tmux.send_keys :Escape, :a
-    tmux.until { |lines| lines[-2].include? '/100' }
+    tmux.until { |lines| lines[-2].include? '/4' }
     tmux.send_keys :Enter
     tmux.prepare
     readonce
-    assert_equal ['["1"], @"1"@', '["1" "2" "3"], @"1" "2" "3"@', '["1" "2" "4"], @"1" "2" "4"@'],
+    assert_equal [%[foo'bar/foo'bar],
+                  %[foo'bar foo"bar foo$bar/foo'bar foo"bar foo$bar],
+                  %[foo'bar foo"bar foobar/foo'bar foo"bar foobar]],
       File.readlines(output).map(&:chomp)
   ensure
     File.unlink output rescue nil
@@ -912,7 +940,7 @@ class TestGoFZF < TestBase
     tmux.until { |lines| lines[-2].include? '1/1' }
     tmux.send_keys 'C-c'
     tmux.prepare
-    assert_equal ['-c / "foo"bar'], File.readlines(output).map(&:chomp)
+    assert_equal ["-c / 'foo'bar"], File.readlines(output).map(&:chomp)
   ensure
     File.unlink output rescue nil
   end
@@ -1026,7 +1054,7 @@ class TestGoFZF < TestBase
     end
   end
 
-  def test_canel
+  def test_cancel
     tmux.send_keys "seq 10 | #{fzf "--bind 2:cancel"}", :Enter
     tmux.until { |lines| lines[-2].include?('10/10') }
     tmux.send_keys '123'
@@ -1163,6 +1191,71 @@ class TestGoFZF < TestBase
     tmux.send_keys :Enter
   end
 
+  def test_jump
+    tmux.send_keys "seq 1000 | #{fzf "--multi --jump-labels 12345 --bind 'ctrl-j:jump'"}", :Enter
+    tmux.until { |lines| lines[-2] == '  1000/1000' }
+    tmux.send_keys 'C-j'
+    tmux.until { |lines| lines[-7] == '5 5' }
+    tmux.until { |lines| lines[-8] == '  6' }
+    tmux.send_keys '5'
+    tmux.until { |lines| lines[-7] == '> 5' }
+    tmux.send_keys :Tab
+    tmux.until { |lines| lines[-7] == ' >5' }
+    tmux.send_keys 'C-j'
+    tmux.until { |lines| lines[-7] == '5>5' }
+    tmux.send_keys '2'
+    tmux.until { |lines| lines[-4] == '> 2' }
+    tmux.send_keys :Tab
+    tmux.until { |lines| lines[-4] == ' >2' }
+    tmux.send_keys 'C-j'
+    tmux.until { |lines| lines[-7] == '5>5' }
+
+    # Press any key other than jump labels to cancel jump
+    tmux.send_keys '6'
+    tmux.until { |lines| lines[-3] == '> 1' }
+    tmux.send_keys :Tab
+    tmux.until { |lines| lines[-3] == '>>1' }
+    tmux.send_keys :Enter
+    assert_equal %w[5 2 1], readonce.split($/)
+  end
+
+  def test_jump_accept
+    tmux.send_keys "seq 1000 | #{fzf "--multi --jump-labels 12345 --bind 'ctrl-j:jump-accept'"}", :Enter
+    tmux.until { |lines| lines[-2] == '  1000/1000' }
+    tmux.send_keys 'C-j'
+    tmux.until { |lines| lines[-7] == '5 5' }
+    tmux.send_keys '3'
+    assert_equal '3', readonce.chomp
+  end
+
+  def test_preview
+    tmux.send_keys %[seq 1000 | sed s/^2$// | #{FZF} --preview 'sleep 0.2; echo {{}-{}}' --bind ?:toggle-preview], :Enter
+    tmux.until { |lines| lines[1].include?(' {1-1}') }
+    tmux.send_keys :Up
+    tmux.until { |lines| lines[1].include?(' {-}') }
+    tmux.send_keys '555'
+    tmux.until { |lines| lines[1].include?(' {555-555}') }
+    tmux.send_keys '?'
+    tmux.until { |lines| !lines[1].include?(' {555-555}') }
+    tmux.send_keys '?'
+    tmux.until { |lines| lines[1].include?(' {555-555}') }
+    tmux.send_keys :BSpace
+    tmux.until { |lines| lines[-2].start_with? '  28/1000' }
+    tmux.send_keys 'foobar'
+    tmux.until { |lines| !lines[1].include?('{') }
+  end
+
+  def test_preview_hidden
+    tmux.send_keys %[seq 1000 | #{FZF} --preview 'echo {{}-{}}' --preview-window down:1:hidden --bind ?:toggle-preview], :Enter
+    tmux.until { |lines| lines[-1] == '>' }
+    tmux.send_keys '?'
+    tmux.until { |lines| lines[-2].include?(' {1-1}') }
+    tmux.send_keys '555'
+    tmux.until { |lines| lines[-2].include?(' {555-555}') }
+    tmux.send_keys '?'
+    tmux.until { |lines| lines[-1] == '> 555' }
+  end
+
 private
   def writelines path, lines
     File.unlink path while File.exists? path
@@ -1229,7 +1322,7 @@ module TestShell
     tmux.send_keys :BTab, :BTab, pane: 1
     tmux.until(1) { |lines| lines[-2].include? '(2)' }
     tmux.send_keys :Enter, pane: 1
-    tmux.until { |lines| lines[-1].include? 'cat' }
+    tmux.until { |lines| lines[-1].include?('cat') || lines[-2].include?('cat') }
     tmux.send_keys :Enter
     tmux.until { |lines| lines[-1].include? 'test1test2' }
   end
@@ -1444,7 +1537,7 @@ module CompletionTest
     tmux.send_keys :BTab, :BTab, pane: 1
     tmux.until(1) { |lines| lines[-2].include? '(2)' }
     tmux.send_keys :Enter, pane: 1
-    tmux.until { |lines| lines[-1].include? 'cat' }
+    tmux.until { |lines| lines[-1].include?('cat') || lines[-2].include?('cat') }
     tmux.send_keys :Enter
     tmux.until { |lines| lines[-1].include? 'test3test4' }
   end